fix(user-server): 修复 Windows 平台无法编译(信号/rusage/statfs 平台拆分)
- endless 依赖 SIGUSR1/2/SIGTSTP(Windows 无此信号),serveHTTP 按平台拆分:
  Unix 走 endless 零停机热重启,Windows 走标准 http.ListenAndServe
- system_monitor 的 getrusage/statfs 为 Unix 专有 syscall,按构建标签拆分:
  Windows 用 GetProcessTimes/GetDiskFreeSpaceEx(golang.org/x/sys/windows)等价实现
- cpuUsagePercent/clampPercent 抽为平台无关公共函数,消除重复
- 验证:windows/linux 双平台 CGO_ENABLED=0 编译通过,
  go vet 通过,service 包 go test 全量通过(偶发 flaky 为上游时序用例,与本次改动无关,
  单独重跑 3 例均 PASS)
